Added Feign client

+++ src/main/java/com/maveric/userservice/controller/UserController.java
@@ -1,7 +1,10 @@
 package com.maveric.userservice.controller;
 
+import com.maveric.userservice.constant.MessageConstant;
 import com.maveric.userservice.dto.UserDto;
 import com.maveric.userservice.dto.UserEmailDto;
+import com.maveric.userservice.exception.UserIdMismatchException;
+import com.maveric.userservice.feignclient.FeignUserService;
 import com.maveric.userservice.service.UserService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -34,6 +37,9 @@ public class UserController {
         this.userService = userService;
     }
 
+    @Autowired
+    FeignUserService feignUserService;
+
     @PostMapping("/users")
     public ResponseEntity<UserDto> createUser(@Valid @RequestBody UserDto userDto) {
         userDto.setPassword(this.bCryptPasswordEncoder.encode(userDto.getPassword()));
@@ -41,8 +47,13 @@ public class UserController {
     }
 
     @PutMapping("/users/{userId}")
-    public ResponseEntity<UserDto> updateUser(@PathVariable("userId") String userId, @Valid @RequestBody UserDto userDto) {
-        return new ResponseEntity<>(userService.updateUser(userDto, userId), HttpStatus.OK);
+    public ResponseEntity<UserDto> updateUser(@PathVariable("userId") String userId, @Valid @RequestBody UserDto userDto,
+                                              @RequestHeader(value = "userid") String headerUserId) {
+        if(headerUserId.equals(userId)) {
+            return new ResponseEntity<>(userService.updateUser(userDto, userId), HttpStatus.OK);
+        }else{
+            throw new UserIdMismatchException(MessageConstant.NOT_AUTHORIZED_USER);
+        }
     }
 
     @GetMapping("/users")
@@ -52,8 +63,13 @@ public class UserController {
     }
 
     @GetMapping("/users/{userId}")
-    public ResponseEntity<UserDto> getUserById(@PathVariable("userId") String id) {
-        return new ResponseEntity<>(userService.getUserById(id), HttpStatus.OK);
+    public ResponseEntity<UserDto> getUserById(@PathVariable("userId") String id,
+                                               @RequestHeader(value = "userid") String headerUserId) {
+        if(headerUserId.equals(id)) {
+            return new ResponseEntity<>(userService.getUserById(id), HttpStatus.OK);
+        }else{
+            throw new UserIdMismatchException(MessageConstant.NOT_AUTHORIZED_USER);
+        }
     }
 
     @GetMapping("/users/getUserByEmail/{emailId}")
@@ -62,8 +78,15 @@ public class UserController {
     }
 
     @DeleteMapping("/users/{userId}")
-    public ResponseEntity<String> deleteUser(@PathVariable("userId") String id){
-        userService.deleteUser(id);
-        return new ResponseEntity<>("User Deleted Successfully", HttpStatus.OK);
+    public ResponseEntity<String> deleteUser(@PathVariable("userId") String id,
+                                             @RequestHeader(value = "userid") String headerUserId){
+        if(headerUserId.equals(id)) {
+            feignUserService.deleteAllAccount(id, headerUserId);
+            userService.deleteUser(id);
+
+            return new ResponseEntity<>("User Deleted Successfully", HttpStatus.OK);
+        }else{
+            throw new UserIdMismatchException(MessageConstant.NOT_AUTHORIZED_USER);
+        }
     }
 }
